Remove TypeSyntax
